requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
slack-sdk>=3.22.0
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0